requirements.txt
# JurisAgent Dependencies
# 法律智能体依赖清单

# ============ 核心依赖 ============
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ============ Web界面 ============
streamlit>=1.38.0

# ============ 文档解析 ============
pypdf2>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# ============ RAG检索增强 ============
chromadb>=0.5.0
numpy>=1.24.0

# ============ 工具库 ============
loguru>=0.7.0
tenacity>=8.5.0

# ============ 开发工具 ============
pytest>=8.0.0
black>=24.0.0
